added some docs + separated the scanner
