+++ lib/packages/watcher.go
@@ -9,25 +9,13 @@ import (
 	"github.com/limetext/lime-backend/lib/watch"
 )
 
-type pkgDir struct {
-	path string
-}
-
-func (p *pkgDir) FileCreated(name string) {
-	record(name)
-}
-
+// Responsible for watching all loaded packages
 var watcher *watch.Watcher
 
+// Helper function for watching a package
 func Watch(pkg Package) {
 	if err := watcher.Watch(pkg.Name(), pkg); err != nil {
 		log.Warn("Couldn't watch %s: %s", pkg.Name(), err)
-	}
-}
-
-func watchDir(p *pkgDir) {
-	if err := watcher.Watch(p.path, p); err != nil {
-		log.Warn("Couldn't watch %s: %s", p.path, err)
 	}
 }
